- Reworked grid-interpolation:
The method Grid::GetElem(s)AtGlobalCoord() is now splitted up in two
parts internally:

  - First we try to determine a set of axis-parallel bounding boxes,
    enclosing the sought-for points. This is done in the method
    Grid::MapPointsToBoundingBoxes(). If CGAL is activated, the fast
    box intersection method is used, otherwise a naive brute-force
    search is applied, which is not depending on any external library.

  - In a second step the method Grid::MapGlobPointsToLoc() is called,
    which iterates over all element candidates associated with a point
    and performs the costly Global2Local mapping and the test, if the
    local point is really within the element.
    Additionally, this method is fully parallelized using OpenMP!

Also, the preprocessor macro USE_INTERPOLATION was completely replaced
by the macro USE_CGAL, which describes more accurately the dependency
behind this switch.

As the naive box-intersection is now the default without CGAL, we can
always perform "interpolation". In addition, the splitting of the
functionality paves the way for other 3rd party libraries, such as
libfbi (a header-only implementation of the fast box intersection
algorithm).

- Moved the interpolation part in the ResultHandler in an own method
  InterpolateRes() for a cleaner layout.
  
- Added use_openmp.hh header, which defines the macro USE_OPENMP for
  future use of openmp system library calls (defined in omp.h).