1) appropriate info for smooth spline approximation is now available
a) cfs with option -a now collects all form the xml-file needed data and
performs the approximation; afterwards cfs finishes; Thereby, one can
quite easily adjust the important paparmeter "measAccuracy" for the smooth-spline
approximation.

b) A nonlinear simulation with smooth spline approximations of provided
data now prints out (in both cases of starting cfs with -a or performing the
full simulation) the following files
Example: bh1.fnc (original file with the provided data)  
- bh1.fncApproxExtrapolate: alpha and beta parameter for the
  extrapolation og the reluctivity(B)-curve towards reluctivity of air
(exponential function beta * exp(-alpha * B)
- bh1.fncApprox: H B (approximated) Bprime
- bh1.fncApproxInv: B H (approximated) Hprime
- bh1.fncApproxNu: B reluctivity reluctivityPrime

2) Smooth spline approximation: uses now in any case the extrapolation
( beta * exp(-alpha * B) ) towards teh reluctivity of air for values of
B larger then provided in the sampled data. So the paramter
"maxApproxVal" provided in the material file is not used. This procedure
guarantees smooth continuation of the reluctivity curve towards larger
B-values.

3) Still the convergence for "hard" nonlinear electromagnetic
computations is not satisfying! Tried a lot!

4) Ansiotropic nonlinear electromagnetic compzutations with a bundle of
BH-curves is now available; definition in the material file is as
follows:

        <nonlinear>
          <anisotropic>
            <data>
              <angle> 0.0 </angle>
              <dependency> magInduction  </dependency>
              <approxType> smoothSplines </approxType>
              <measAccuracy> 0.01 </measAccuracy>
              <maxApproxVal> 2.0 </maxApproxVal>
              <dataName> bh_wr.fnc </dataName>
            </data>
            <data>
              <angle> 90.0 </angle>
              <dependency> magInduction  </dependency>
              <approxType> smoothSplines </approxType>
              <measAccuracy> 0.01 </measAccuracy>
              <maxApproxVal> 2.0 </maxApproxVal>
              <dataName> bh_qr.fnc </dataName>
            </data>
          </anisotropic>
        </nonlinear>

Currently, as for the magnetostrictive strains, we just consider the x-
and y-component of the magnetic flux B and compute from this the angle!!