commented out Global2LocalCoords --
The new Global2LocalCoords(...) does not work with non-matching.
Reason is, that it can not handle 2D Elements which are in 3-Dimension.
So surfaces on a volume as for non-matching grid loose there third
coordinate.

I just commented it out so the old Global2LocalCoords is active. But the
new one is much better an more reliable. Needs to be fixed so it would
work for 3D.