This reenables the CoefFunctions Grid. To reduce computational time and memory requirements in the interpolation process, the newly defined CoefFunctionNodalGrid is specifically implemented for nodal results. As hdf5 is not capable to store higher order results, this should be fine for now. In future we can reactivate the original CoefFunction which is prepared for higher order results.
The grid coeffunction has now everything to interpolate the results also
in the time domain which will be used in furture applications when
conservative interpolation is moved to the FeSpace.

The user can now specify a factor along in the grid tag which
enables multiplication of the ceofVector by a time or space dependent
factor.

Added two additional functions to the elementShapeMap Global2Local
procedure. The first one checks if the point coincides with a corner
node, the second one is the straight forward implementation which can be
found in the scriptum by Duester. Up to the current experience, these functions give for QUAD and HEXA
elements better results even in the quadratic case. If we experience
further problems with this we should switch back to the general version.