clean up fluid->mech coupling --
The shear viscous stress and pressure force are now calculated exact.
The changes work but the impact is still to be analysed---ongoing.
No significant loss in speed is detected (1%).

The new method required the new Global2Local method implemented by ahueppe.
Since this caused problems with the non-matching algorithm in 3D the new
method is only called if it is feasible.
The reason is that the method only works in pure 2D. In other words,
the method did not work for surfaces which have a z-Component != 0.


fluidMechShearStress.*
  - correct calculation of the surface stress tensor


fluidMechPDE.cc
  - gathering of element matrices for coupling
  - clean up of code

basefe.*
  - check if new fastGlobal2Local() method may be used

quad2fe.*
  - included fastGlobal2Local() method