This check-in circumvents a problem with Ensight Data obtained by FOAMPro but does not ultimately resolves it.
The problem is due to the fact, that different CFD codes or their Preprocessors permutate their element
connectivity. Due to this, one has to reorder it to ensure positive jacobians. Right now, the solution is to
put the corrected ordering directly in the FileReaderEnsight class rather than its base. Hopefully this dows not break any
other code. To completely resolve this issure, one would need to check for negative jacobians when the grid is loaded.