Changed the default rule for the integration order: Previously it was calculated as
	integOrder = 2 * (p+1)

which in case of quadrilateral elements of 1st order led to 9
integration points (classical gauss integration). In order to comply
with the trunk version, we now choose the rule

	integOrder = 2 * (p+1) - 1

leading to integOrder 3 and 4 integration points for the first order
quadrilateral. This rule should be
valid also for higher order polynomials and increases the
performance.