Implemented restart functionality:
It is now possible for transient simulations to halt them using
Ctrl-C. In this case the user is informed,  the current time step 
is finished and in the end restart information is written 
for the current step to the HDF5 file. 
The simulation can be re-run by calling CFS with the -r argument.

Also, we write by default restart information for the last time step,
so if the number of time steps is increased in the xml file after
a full simulation, additional time steps can be simulated. Again,
the -r switch has to be used. If no restart information should be 
written, the <writeRestart> element of the <transient>-section 
should be set to "no".

The restart functionality is supposed to work for all single and
coupled field problems, even iterative ones, so no need anymore
for complicated Memento and CouplingMemento objects.

At the moment, only the HDF5 writer supports the restart-functionality
in a clean way, i.e. we can extend the simulation arbitrarily often,
as all arrays are now internally extended step by step instead of
creating them for a fixed size.

Additional internal changes:

- Added new methods in HDF5IO class allowing to gradually extend
  1D and 2D arrays.

- Added commodity function in HDF5IO which tries to open a group
  or otherwise creates it.

- All output writers get now the information, if the current run
  is a restarted one, i.e. the result files should not be
  overwritten (if possible). Currently this information is only
  used in the HDF5 writer.

- Removed debugging output and the adjointParams arguments.