Refactored the mapping of virtual nodes in the FeSpace classes.
Previously the FeSpace stored the virtual nodes for
vertices/edges/faces/interior per element, as in the most general
case the mapping is discontinuous (L2 space) and thus there is no
unique global numbering of vertices/edges/faces. Therefore it was not
possible to access virtual node numbers (and equations numbers) for
elements, which were not previously known to the space (e.g. due to
boundary conditions or RHS integrators).

This leads to problems if e.g. we query the equation numbers for
surface elements, which are not explicitly known in the space. 

However in most cases we have continuous spaces (H1, HCurl, HDiv), so
there is now an alternative mapping, which only stores the virtual
nodes globally per node/edge/face/element. Thus, we save memory (as
each entity gets only stored once) and we can also query virtual
nodes / equation numbers for arbitrary elements by taking their
vertex/edge/face numbers and look in the corresponding global array.

One drawback is, that in case of nodal permutation (as used in the
Lagrangian elements of higher order) we have to apply the mapping each
time the equations are queried, but this should not pose too much
additional work. Previously this mapping was stored already for each
element.

The old, element based node mapping is still used for the FeSpaceL2,
which does not use the new data structures at all.

- Added new method BaseFE::NeedsNodalPermutation(), which is only true
  for elements requiring a true permutation (i.e. the Lagrange
  elements of variable order).

- Fixed some bugs in the method FeH1LagrangeVar::GetNumFncs, 
  which set wrong entries.

- Adjusted FeSpace classes to new mapping structure