Extended calculation scheme of magnetic nonlinear and anisotropic materials:
So far, the calculation of anisotropy was only performed within the xy plane of material.
-> Added consideration of z orientation of magnetic flux to evaluate the underlying nonlin curves.

Note: 
Since we do not yet have nonlinear curves in z direction we use the same curves as given 
for the xy-plane but scale it with an appropriate factor.

First tests with simple model successful; detailled comparison of impact still has to be done...