THis enables the Nitsche method for rotating Sytems and works for acoustics. The Moving mortar flag is currently used to Switch from penalization of the quantity itself to the penalization of the first time derivative. Tests are currently running in determining the correct penalization factor. The current situation reveals the following choices: < 100  static Nitsche 2D 1-10   rotating Nitsche 2D < 1    static Nitsche 3D << 1   rotating Nitsche 3D
More investigations necessary to determine the reason for this cascade. 
It is still not clear if this behavior is due to the method or due to flaws in the implementation.

For the preperation of multiple rotating interfaces, the reset of the interface (deletion of nodes and elements) in each time step is performed in a seperate step before performing the update. This is still not the clean solution and an excetpion in GridCFS.cc will prevent the deletion of nodes which are not at the end of the node list. Disabling this exception will make multiple interfaces work but is very sensible to errors.