Implemented concentrated mechanical network elements (spring, mass, damping element):
The previous implementation of the "spring" command did only work for
nodelists with 1 node. The new implementation can handle "point"
masses/stiffnesses/damping values, as well as pairwise values (e.g.
spring / truss element), where the nodelist consists of 2 nodes. If
the nodelist contains more than 2 nodes, a warning is issued and the
element is not considered.

To reflect the fact that we model in general concentrated network
elements rather than mechanical springs, the element <spring>
was renamed to <concentratedElem>.