Resolves a serious but (at leat for gcc 4.8.1:
At least in CoefFunctionFormBased instead of Conj<Vector<Complex> >() from Vector.cc
it was used
 template <typename T>
 inline T Conj( T a ) {
   return a; 
 }

from opsdefs.hh in release mode. Therefore the conjugate of a complex vector was the vector itself.
In debug mode the right Conj() was used.

As solution there is now a double conj() and a complex conj() in opsdefs.hh.

std::conj() cannot be used, as double, int, ... is in the C++ standard only sice C++11 :(

Same with boost::conj() - it is in boost but not in CFS++.

By the way, instead of using Conj() for the vector used Vector<T>::Inner(), this requires no temporary object.