This implements some of the performance tweaks, Andi Hauck introduced to NACS Mainly, this covers the following: - Element shape map gets only created if it is really needed   To cope with this, we use the reference counting mechanism of boost shared ptr   and create these structures spereately for each omp thread which may occure   thereby we prepare for parallel matrix assembly.   Currently, the Method is a bit circumenfencial as we also need to cover the   mortar interface case in which the element does not necessarily have an unique   element number but 0 - Derivative of element shape functions are cached by thread - equation numbers are chached by thread - Algebraic system is tweaked regarding the insert matrix method - StdVector Clear can avoid freeing memory by boolean switch
To sum up, due to these modifications which are a combination of caching and 
avoiding millions of allocates and frees saves already some time when running the 
testsuite and we make some more steps w.r.t. to parallelization of the assemble loop.

In addition, the EXCEPTION "A line cannot cross than 2 edges of a convex polygon" has 
been chanded to a warning. This case simply occurs too often on free interface meshes.
Modifications to NCInterface class regarding CGAL intersection algorithms, tree search 
to avoid N_masterElem*N_slaveElem intersection complexity as well as avoiding
the "on demand" insertion of interseciton elements into the grid are pending.