Fixed/Commented some issues with latest commit concerning inductance. Here are my thoughts about this topic, since it is not an easy one and there are still some issues in cfs.
1) The result "coilInductance" will not work for voltage coils. The reason
is that the coilCurrentDensity currently cannot be calculated correctly for
voltage coils. This is related to FeFunctions and the FeSpaceConst and
requires some effort to fix. The coilCurrentDensity for voltage coils would
also require the knowledge of the coil cross section. The automatic
calculation of the cross section is currently not implemented. I added
comments at the certain positions in the code.

2) The implemented "coilInductance" is calculated by Psi/i (flux linkage by
current). This is commonly called the static inductance, secant inductance or
large-signal inductance.
2a) For non-linear material, this calculation still works but it is a matter
of definition.
2b) As soon as a permanent magnet is involved, it is wrong since the flux is
not necessarily zero if the current is zero.
2c) In my opinion, the more general form of the inductance is the differential
inductance. It is calculated as the derivative of Psi by i. For a linear setup,
it is the same as the static inductance, because the Psi versus i diagram is a
straight line through the origin. For non-linear setups including permanent
magnets, it is the correct inductance in terms of an electric circuit.
However, we cannot compute it in static simulations directly in cfs because it
needs a change of i. Personally, I always calculate the inductance with the
help of Matlab or Python, since cfs can write Psi and i as a text result.

3) If calculating quantities like energy, flux or inductance, the voltage coil
behaves in quite a special way. Depending on the symmetries in the geometry,
different scaling factors have to be applied to voltage, wire cross section,
resistance, current, induced voltage and flux linkage. As an example of the
verification of these quanities, the test suite contains a 3D example of a
voltage loaded coil, verified by the time constant of the R-L-series.
In contrast, current loaded coils are only affected in the post-processing,
e.g. multiply energy by 2 if 1 symmetry plane is used in 2d.

For those reasons, I did not implement "the" general inductance as a result so
far.