Made some changes to cfs-debug script (still work in progress and may break MacOS compatibility?).
IMPORTANT for every user:
 If you experience problems with a run of the CFS binary (segmenation faults, exceptions, etc.) do
 the following before you report the error:
 1. Execute the command "cfs-debug gdb -f <arguments>".
    The script "cfs-debug" is also located in the "bin" directory of your CFS installation 
    Replace <arguments> with the exact cfs command line arguments which led to the error
 2. Follow the instructions given at the end of the script about the important files to include in
    the report
 3. If you have access, use the ticket system https://cfs.mdmt.tuwien.ac.at/trac/newticket
    and include the requested files

A version of the debug script for cfsdat is not yet available.

This checkin also fixes some issues with the openMP related TLMap and the isCleared_ flag
In case of stl maps with TLMap isCleared_ was set to false by standard constructor which causes
assertion in destructor to fail on immediate override e.g. by assignement operator.
Maybe, assignement operator will be removed and replaced by a Reset function in future releases.

Also, this commit should fix some issues with the vtk compile on CentOS which arose due to an
error in the semicolon patch in cfsdeps/vtk/vtk-semicolon.patch