This is the first attempt to fix multiple output filters and branched pipelines in general.
The output filter now deactivates its upstream results which should be valid in general as 
an output is always the end of the pipeline. 
Additionally, in extract filters, additional results can be added without cleaing the filterResId 
vector first.

NOTE:
Multiple output filters should work now although the current concept of the run methods in the filters
seems to turn out to be too error prone. Everything is implemented inside the run method of each filter
and it is quite easy to make mistakes. For future developments, the structure of the run methods 
might need refactoring to ensure a more failsafe and generic implementation.