fix for GCC and MKL2017
By default the OMP-lib is linked dynamically resulting in 'libiomp5.so not found'
when running CFS. This could possibly be resolved by creating a custom target to
copy over the '*.so', or statically link libiomp5. We link statically, although
it is not recommended by intel (anylone knows why?).

The script 'compile_mkl_test.sh.in' needed some adaptions:
 - OMP_LIB_PATH is now determined by seraching for 'intel64_lin/*iomp5.a'
 - replace '-liomp5' by absolute path to library 'libiomp5.a' to link statically
 - some debug output

In order for this changes to take effect you need to remove MKL_BLAS_LIB and MKL_ROOT_DIR
from your CMakeCache!

Hopefully, this version still works with older MKL-versions.