new structure for BaseEigenSolver and more cases for FEAST
FEAST
---------------------
can now handle real-general, real-symmetric and complex-hermitian
and uses the correct xml input to specify a search interval
      <eigenFrequency>
        <isQuadratic>no</isQuadratic>
        <minVal>0</minVal>
        <maxVal>57</maxVal>
      </eigenFrequency>
See the testsuite for examples.

BaseEigenSolver
--------------------
The BaseEigenSolver has a new structure to provide a unified interface to different
eigenvalue solvers:

The init methods
  Setup(const BaseMatrix & A, bool isHermitian=false) -> for standard EVP
  Setup(const BaseMatrix & A, const BaseMatrix & B, bool isHermitian=false) -> for generalised EVP
analyse the matrix properties and set the new EigenValueProblemType to
  REAL_SYMMETRIC, REAL_GENERAL, COMPLEX_SYMMETRIC, COMPLEX_HERMINTIAN, or COMPLEX_GERNERAL

The new query methods
  bool HasComplexEigenvalues(), and
  bool HasComplexModes()
then can be used to determine the type of eigenvalues and modes.

The new computation methods
  void CalcEigenValues( BaseVector &sol, BaseVector &err, Double minVal, Double maxVal )
  void CalcEigenValues( BaseVector &sol, BaseVector &err, UInt N, Double shiftPoint )
should be used to obtain eigenvalues from the solver.

Eigenvectors can be obtained by the old method
  GetEigenMode( UInt modeNr, Vector<Complex> & mode )
This should be changed in the future to have the possibility to extract left and right eigen modes.

The case of the quadratic EVP was not adressed deliberately. Any Eigensolver that can handle general matrices
can be used to solve a quadratic EVP. As the procedure is the same for any solver, it does not make sense to
re-implment it for all of them. The plan is to move this functionalty to the Driver.

All intermediate methods in AlgebraicSys and StdSolveStep should eventually be removed.
When using FEAST this is already acomplished: All necessary Setup is done in EigenFrequencyDriver.
This is done by simply getting the pointers for AlgSys and BaseEigenSolver.

This new interface currently is only activated if <minVal> and <maxVal> are used in the input.

Sill Missing
-----------------------
 - pretty output of eigenfrequencies to terminal
 - sorting of modes (currently only done for the output)
 - export of system matrices: just call the method from AlgSys
 - make USE_FEAST=ON the default
 - fast stochastic estimate for the number of EVs in an interval using FEAST -> could be used to get a specified number of EVs