Two new compare modes for cfstool and config options for testsuite
The modes are called 'L2diff' and 'relL2diff' and are invoked equivalently to the
original mode 'scalardiff' used in the testsuite runs, i.e.
$ cfstool -m ${CFS_TESTSUITE_CFSTOOL_MODE} --eps ${EPSILON} ${H5REF_FILE} ${H5_FILE}"

Both modes are based on the L2 norm of the data vectors.
The norm is evaluated independently for each result on each region
for each time/frequency step and for each MultiSequenceStep
(for mesh and history results) as in the 'scalardiff' mode.
This is done in the new function `CheckL2` of `cfstool.cc`.
It always evaluates the L2 norm of the differnce vector, i.e.
 L2(v-v_ref)
and the relative (=normalised) L2 norm or residual, i.e.
 L2(v-v_ref)/L2(v_ref)
Addtionally data like minimum, maximum, and average values of
the data in the difference and reference vectors are output.
The output will also tell you where exactly (MultiSequenceStep, Step,
result and region) the largest residual is encountered!

The tolerance check dependes on the mode:
 'L2diff' checks against both absolute and relative L2 norm
 'relL2diff' checks against relative L2 norm only (thus, more robust)

I propose to switch to 'relL2diff' as the default mode in the future
because of the following advantages:
 + the code in `CheckL2` is much more transparent
 + equivalent handling of all analysis types
 + useful debug output on difference
Please review `cfstool.cc` and test the new mode on your machine.

The new config options for the testsuite are:
CFS_TESTSUITE_CFSTOOL_MODE=scalardiff|L2diff|relL2diff
CFS_TESTSUITE_THREADS=n (integer)
and can be used to control the testsuite behaviour.
The first one defaults to 'scalardiff' and thus reproduces
the original behaviour. The second one defaults to the
number of CPUs of the machine, and is used to set the value
of the '-t' argument for the cfs runs in the testing procedure.

Associated testsuite commit will follow.

At least on my machine the mode 'relL2diff' is more robust than 'scalardiff'.
Apart from cfs-segfaults the following tests fail:
	 45 - Singlefield_Electrostatics_MovingMortar (Failed) -> only step 29 is completely wrong.
	130 - Coupledfield_MagStrict_MagStrictMovingMortarMeshUpdate (Failed) 3.13706e-06>tol
	132 - Coupledfield_ElecTherm_ElTh_NMGtransient (Failed) 4.15722e-06>tol