added info output to cfstool
This makes it possible to get the "error" value of a test in CDash!

CTest will automatically parese everything of the form
  <DartMeasurement name="pi" type="numeric/double">3.14</DartMeasurement>
issued by the test command in stdout. The field "pi" is then availabe in CTest
as a data field like "Status" or "Time" and can, thus, be displayed in CDash.

One could use the same for various other test metrics like times in performance testing ...