added structure for averaged region results
We use ResultFunctorIntegrate to obtain the integral of an
arbitrary CoefFunction over one region. When setting `average_`
to true, the integral is devided by the region volume.

Thic can be used for computing
 - homogenised results
 - center of mass
 - center of volume
 - centroid of areas
 - reaction moment
 - ...

Currently implemnted in MechPDE for stress and strain, i.e.
the new region results 'mechStressAverage' and 'mechStrainAverage'.

One limitation is that we cannot combine two regions, because
we lack the necessary structure, however, this is easily done
in a postprocessing step.

C++ Tipp: When you want to dynamic_cast a shared_pointer ptr you need ptr.get() as the argument of the cast.