new material parameter reading for mechanics (isotropic)
This approach is base on the new CoefFunction based structure of
BaseMaterial, that was only partly used in mechanics. It works
as follows:
1) XMLMaterialHandler reads in strings from the xml,
2) transforms them into strings for the stiffness tensor
3) saves the tensor to a CoefFunction via `material->SetCoefFct(MECH_STIFFNESS_TENSOR,C)`

No more Finalize() in the MechanicMaterial required. I suppose
most of the stuff in MechanicMaterial can be deleted, because
it is obsolete.

PS: This also fixes a bug - non-constant E or nu were simply not working ...