added left eigenvectors (modes) to interface in BaseEigenSolver
This way it is possible to choose which eigenvectors are returend by GetEigenMode.
By default the right EVs are returned (this is the original behaviour).

There is currently no XML option to select which modes to return.
The left modes can be obtained by exportLinSys. Currently only
when using FEAST.