enabled the assembly of a complex valued EVP
This naturally occurs when using complex valued
material parameter (e.g. viscoelasticity in mechanics
or viscosity in acoustics).
We simply check if any system matrix is complex, and if
yes, convert all matrices to complex as it is done in a
harmonic analysis.

Additionally: enabled coupled mechanic-acoustic eigenvalue
analysis in the pressure formulation - works with FEAST.