make cfstool work for general complex modes
Only implemented for the relL2diff and L2diff modes!

The Eigenvectors have an arbitrary phase and amplitude.
Therefore, the "sign-switch" did not work (e.g. for Bloch, where phase is random).
We need to normlize the, in general, complex mode by:
1) search for the position of the maximum absolute value of the reference Eigenvector
2) rotate the entire eigenvector such that it has phase 0 and amplitude 1 at this location
   i.e. divide the vector by the maximum absolute element
3) do the same for the same position in the second mode vector
This is the generalized version of the "sign-switch".

Additional methods in the `Vector` class:
MAC - the modal assurance criterion for comparing two complex valued mode shapes
MaxAbs - return value (and position) of the element with maximum absolute value

PS: In real-general EVPs the normalization was not reproducible with
the FEAST eigenvalue solver.