integrated the testsuite as a git submodule
This way it is automatically checked out, and it is possible to define
its revision together with CFS.
The TESTSUITE_DIR is automatically defined to point to this submodule,
and is now an advanced option (hidden by default).

The option CFSTOOL was renamed to BUILD_CFSTOOL.

Cmake will issue an error if incompatible options are selected.