normalSurfaceMass boundary condition in LinFlowPDE
It adds an inertia term for movement in normal direction of boundary surfaces.
This is accomplished by a bilinear form \int v'.n m v.n ds on the surface,
where the parameter m represents a mass per area. It can be used to
"constrain" normal motion to zero on the boundary if the parameter m is
"high enough", i.e. it becomes a penalty parameter.
This is useful if the boundary is not aligned to coordinate axes (where
Dirichlet conditions can be used), or not plane.

define it as

<normalSurfaceMass name="L_top" value="100" volumeRegion="S_flow"/>