remove mingw, cygwin and crosscompiling
crosscompiling (win on linux, mac on linux) is not required any more as
we can now do everything native.

mingw complicated the already too complicated cmake stuff and was not
used by a long time (if ever for real usage).

Note that already on the current state, parts of mingw might already
have been removed!

Probably one wants to add the gfortran compiler on windows again, but
then much less code shall be necessary.

There is a matching testsuite commit with some minors