Finetuning and correction of Mayergoyz and Sutor Vector Hysteresis models I. Mayergoyz model 1. Added three additional parameters to mat.xml for fine-tuning the Mayergoyz Vector Hysteresis model a) useAbsoluteValueOfdPhi b) normalizeXInExponentOfG c) restrictionOfPsi A detailed description of these parameters and their function is documented directly in the xml-schemein form of an annotation. Thus it is directly accessible by the user. Note that these parameters were already present in the hysteresis model but until now they were hardcoded. However, the default value of useAbsoluteValueOfdPhi is now false (formerly true) as tests have shown that the absolute value of dPhi in the computation of the additional lag correction only suits for the rotation in counterclockwise rotation but not in clockwise fashion.
II. Sutor model
Improvement of local inversion by allowing bisection for revised Sutor model.
By explointing the fact that the applied field and the output of the hysteresis
operator aligns for the input's amplitude exceeding inputSaturation/rotationalResistance
for rotationalResistance < 1, the revised Sutor model can be inverted for such
amplitudes using a bisection of the anhysteretic contribution only. This
provides not only a serious acceleration in terms of runtime but further
improves the invertibility.
Prior to this commit, a solution via bisection was only allowed for the scalar model and the
classical variant where such an alignment is already guaranteed for the input
to reach inputSaturation. Note in this context, that the Mayergoyz model cannot
utilize bisection as here input and output of the hystersis operator are not
guaranteed to align.

III. Further extensions and modifications related to hysteresis models
1. Extended computation of InnerAngle function for the 3d case.
As the extended Mayergoyz vector hysteresis model requires the evaluation of
projection and lag-correction angles, one has to ensure that both are measured
in the same angular sense. In 2d this is no issue as the reference axis for
rotation is fixed to the positive z-axis. In 3d however only the absolute
value of the angle is retrieved from the general formula and in exchange
the rotation axis swaps sign. A simple, but unfortunately not always working
approach for receiving a common reference axis for all angle computations is
to ensure that one of the components (here the z-axis) should always have the
same sign. By enforcing that the z-component of the rotation axis (retrievable
from the crossproduct of the two vectors) is positive, one obtains the same
positive/negative angles between two vectors inside the x-y-plane as when
computing the angle from the x-y-subvectors.
A corresponding testroutine was added to the Mayergoyz class (as it is currently
the only class that utilizes the innerAngle function).
2. Edited the nonlinear logging information and extended the statistics
that are produced during the testing routines for the hysteresis models.
3. Nan prevention for local FP-inversion and bisection (strong divergence
on element level will be directly detected and leads to simulation stop;
otherwise cfs would run further and further but producing only Nan results)