Interpolation of material catalogues in DesignMaterial.
Move inline keyword in DesignMaterial to function definitions.

Add function to generate vectors with random values in Utils/tools.

Changes in interpolation classes from Utils/ApproxData:

* Add unit test for cubic interpolation (cubicinttests.cc).

* Fix bugs in *CubicInterpolate, especially for periodic data.

* Remove unnecessary code.

Improve interpolation of material catalogues:

* Change interpolation in DesignMaterial from custom implementation in this class to the one in ApproxData.

* Change volume interpolation in Function from custom implementation in this class to the one in ApproxData.

* Change monocubic interpolation in matlab:
For cubic interpolation, the value of the derivative at the boundary has to be given. This was previously done by forward/backward differences. This commit changes the interpolation to be quadratic in the boundary pieces of the interpolation space. No derivative has to be given at the boundary and the resulting polynomial is still C1.

* Remove duplicated scripts: Move all matlab interpolation scripts to matlab code folder and link in python code folder.

Fix bug in derivatives of SplineBoxDesign.