Added L2Norm postproc result
- Added the option to calculate the absolute or relative L2 norm of an arbitrary result with respect to the analytical reference solution
- Here, we actually evaluate the function and not just the l2 norm of the nodal values. Hence, an additional integration order needs to be chosen high enough to accurately integrate the analytical solution
- Fixed a bug where the post processing results (xml) got read in from the previous sequence step instead of the current one
  * Now, the current sequence step gets set twice, but removing the previously used assignment of the sequence step breaks a lot of tests
- Added two testcases
  * One for scalar quantities (heat) and one for vectorial (smooth)
  * We are also testing absolute and relative norm as well as different input and integration orders

Usage guide:
Currently only scalar and vectorial fields are implemented.
For scalar problems, the name of the DOF should be left empty, for vectorial problems, each DOF has to be set with the correct name (e.g. "x", "y" and "z").
Depending on the type of the problem (real valued or complex) one has to proved realFunc only or realFunc and imagFunc for complex fields.
For the calculation either an absolute or a relative is possible. Regarding the integration order one has to choose the order high enough to get an accurately intgrated analytical result.