Adapted LinFlowPDE coeff func. for symmetric case
Throughout the LinFlowPDE class the coefficient
functions of the BLF integrators of balance of
momentum and of all LF integrators have been
changed in case the symmetric formulation is used.
The coefficient functions have been multiplied by
a constant factor minus one.

Note: Testing of the symmetric case is not yet
possible, so only the asymmetric case has been
tested!