Adapted LinFlowHeatCoupling coeff func. for symmetric case
In the LinFlowHeatCoupling class the coefficient
function of the flow to heat integrator has been
changed in case the symmetric formulation is used.
The coefficient function has been divided by the
reference temperature.

Note: Testing of the symmetric case is not yet
possible, so only the asymmetric case has been
tested!