implemented external solver
The external Solver exports the system matrix and rhs of the to be solved system and calls another program via the terminal, which solves the system.
The external command is supposed to write the solution vector to disk.
Afterwards the solution vector is imported into CFS.

The argument list for the external command is user defined via the XML input.
Special arguments `<rhsFileName>`, `<matrixFileName>` and `<solutionFileName>`
define the IO-interface between CFS and external program.
If they are not given, the files are not exported. This might be beneficial if
no export is needed because the solution can be provided by other means.

Special argument `timeFreq` can be used to output the time/frequency value.

Usage example
```
    <linearSystems>
      <system>
        <solverList>
          <externalSolver>
            <logging>no</logging>
            <cmd>@PYTHON_EXECUTABLE@ solve.py</cmd>
            <arguments>
              <rhsFileName>rhs_vec</rhsFileName>
              <matrixFileName>system_mat</matrixFileName>
              <arg>-s</arg>
              <solutionFileName>sol_vec</solutionFileName>
            </arguments>
          </externalSolver>
        </solverList>
      </system>
    </linearSystems>
```
see testsuite for further usage examples.

Signed-off-by: Florian Toth <florian.toth@tuwien.ac.at>