Implement External Eigensolver
The external eigensolver allows for the usage of an external script (that must be callable via the command line) for the computation of eigenvalues.
The Matrix Market exchange format is used for the export of the system matrices from openCFS and import of the result vectors for eigenvalues and eigenvectors from the external script.
In the XML scheme, the solver call as well as a certain number of arguments can be specified.
These arguments will be added to the command in their order of appearance in the <arguments> section.
The external eigensolver can be defined as follows:
<eigenSolverList>
  <external id="ext">
    <logging>yes</logging>
    <cmd>python3 EigenSolver.py</cmd>
    <arguments>
      <AFileName/>
      <shiftPoint/>
      <EigenValuesFileName/>
      <EigenVectorsFileName/>
    </arguments>
  </external>

Also remove exception for quadratic eigensolver since it can now work without cfsdeps.

Testcases for external eigensolver